% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize_first}
\alias{capitalize_first}
\title{Capitalize first letter of a string}
\usage{
capitalize_first(text)
}
\arguments{
\item{text}{String to capitalize}
}
\value{
String with first letter capitalized
}
\description{
Capitalize first letter of a string
}
