% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_config}
\alias{default_config}
\title{Default survey configuration}
\usage{
default_config(
  tries = 2,
  response_delay = 0,
  character_delay = 0.02,
  delay_variance = 0.01,
  version = "1.0"
)
}
\arguments{
\item{tries}{Maximum retry attempts for unclear responses}

\item{response_delay}{Delay before bot response (seconds)}

\item{character_delay}{Delay between characters in streaming (seconds)}

\item{delay_variance}{Randomness factor for character delay (seconds)}

\item{version}{Survey version}
}
\value{
Configuration list
}
\description{
Default survey configuration
}
