% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variable names from template strings}
\usage{
extract_variables(template)
}
\arguments{
\item{template}{Template string with {variable} placeholders}
}
\value{
Character vector of variable names
}
\description{
Extract variable names from template strings
}
