% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{increment_retry}
\alias{increment_retry}
\title{Increment retry count for session}
\usage{
increment_retry(con, session_id)
}
\arguments{
\item{con}{Database connection}

\item{session_id}{Integer session ID}
}
\description{
Increment retry count for session
}
