% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialect.R
\name{on_connect}
\alias{on_connect}
\title{Run the statements a driver needs immediately after connecting}
\usage{
on_connect(con)
}
\arguments{
\item{con}{Database connection}
}
\description{
Run the statements a driver needs immediately after connecting
}
