% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{personalize_text}
\alias{personalize_text}
\title{Personalize question text with user responses}
\usage{
personalize_text(text, responses)
}
\arguments{
\item{text}{Question text with {placeholders}}

\item{responses}{List of response values}
}
\value{
Personalized text
}
\description{
Personalize question text with user responses
}
