% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{start_session}
\alias{start_session}
\title{Start a new survey session}
\usage{
start_session(con, version = "1.0")
}
\arguments{
\item{con}{Database connection}

\item{version}{Question set version}
}
\value{
session_id
}
\description{
Start a new survey session
}
