% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{survey_ui}
\alias{survey_ui}
\title{Chat survey page}
\usage{
survey_ui(
  title = "SurveyChat",
  chat_id = "chat",
  progress_id = "survey_progress",
  footer_id = "survey_footer"
)
}
\arguments{
\item{title}{Card header title}

\item{chat_id}{ID of the chat element}

\item{progress_id}{Output ID of the progress cue}

\item{footer_id}{Output ID of the completion footer}
}
\value{
A Shiny UI definition
}
\description{
Chat survey page
}
